common: use safer uint64_t for list size

+++ src/include/xlist.h
@@ -63,7 +63,7 @@ public:
 
 private:
   item *_front, *_back;
-  int _size;
+  size_t _size;
 
 public:
   xlist(const xlist& other) {
@@ -79,7 +79,7 @@ public:
     assert(_back == 0);
   }
 
-  int size() const {
+  size_t size() const {
     assert((bool)_front == (bool)_size);
     return _size;
   }
